Only add the gmake symlink when it is missing

+++ packages/devel/make/package.mk
@@ -35,5 +35,7 @@ PKG_AUTORECONF="no"
 export CC=$LOCAL_CC
 
 post_makeinstall_host() {
-  ln -s make "${ROOT}/${TOOLCHAIN}/bin/gmake"
+  if [ ! -f "${ROOT}/${TOOLCHAIN}/bin/gmake" ]; then
+    ln -s make "${ROOT}/${TOOLCHAIN}/bin/gmake"
+  fi
 }
